% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_view.R
\name{close_view}
\alias{close_view}
\title{Close All View Tabs in RStudio}
\usage{
close_view()
}
\value{
None
}
\description{
This function closes all view tabs in RStudio. It starts from the current document and moves to the next tab until it reaches the starting document again.
}
\details{
Note that this function will prompt you to save any unsaved files. If you choose not to save, these files will be closed without saving.
}
\note{
This function requires the \code{rstudioapi} package, which is not automatically included in R. You need to install it manually if you haven't done so.
}
