% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_info.R
\name{get_file_info}
\alias{get_file_info}
\title{Get Column Names and Row Count}
\usage{
get_file_info(filepath, verbose = FALSE)
}
\arguments{
\item{filepath}{The path to the file.}

\item{verbose}{If \code{TRUE}, message the number of columns and rows.}
}
\value{
A list with the number of columns and rows.
}
\description{
This function returns the number of columns and the number of rows in a given file.
It uses the \code{system} function to execute shell commands.
}
