% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_git_branch.R
\name{get_git_branch}
\alias{get_git_branch}
\title{Get Current Git Branch}
\usage{
get_git_branch()
}
\value{
A character string representing the current Git branch.
}
\description{
This function returns the name of the current Git branch.
It uses the \code{system} function to execute the \verb{git branch --show-current} command.
}
