% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_operators.R
\name{\%=\%}
\alias{\%=\%}
\title{Exact match operator}
\usage{
a \%=\% b
}
\arguments{
\item{a}{The first value to compare.}

\item{b}{The second value to compare.}
}
\value{
A logical value indicating if the values are identical.
}
\description{
This operator checks for exact matches between two values in R.
}
\examples{
1 \%=\% "1"  # FALSE
1 \%=\% 1    # TRUE
}
