% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Branch.R
\name{sa_branch_set}
\alias{sa_branch_set}
\title{SA Branch set}
\usage{
sa_branch_set(branch)
}
\arguments{
\item{branch}{String object with name of the branch}
}
\description{
Sets a global option determining the current branch
}
