% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.R
\name{sa_network_dir_set}
\alias{sa_network_dir_set}
\title{SA Network dir set}
\usage{
sa_network_dir_set(branch)
}
\arguments{
\item{branch}{String containing the name of the branch}
}
\description{
Sets global option determining the current network directory
}
