% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{src_clean}
\alias{src_clean}
\title{Source clean}
\usage{
src_clean(file, ...)
}
\arguments{
\item{file}{path to script.}

\item{...}{optional parameters to parse to base::source()}
}
\description{
Source script without saving objct to global environment.
}
