% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_view.R
\name{close_view}
\alias{close_view}
\title{close_view.}
\usage{
close_view()
}
\description{
Close all view tabs.
Note: this also closes unsaved files.
You will be prompted to save before these files are closed.
}
